[XEB] Generalize cirq.google.calibration for XEB (#3881)
Add analogous classes for XEB calibration. This re-configures some of the plumbing to be more general. Most of the structure of the Floquet objects should be untouched / compatible. 

This is a proof of concept that I wanted to get initial comments on from @mrwojtek and @dstrain115 before committing and putting on all the polish (generalize the rest of the functions, tests, notebooks)

Open questions:
 - should the PhasedFSimCalibrationRequest objects have an attribute which is the class of request that they are associated with? 
 - Do you want to go through a deprecation cycle for `prepare_floquet_characterization_for_moment(s)`?