Provide sweep-iterator API. (#4094)
Fixes #4064.

This PR provides `*_iter` methods for all simulator "sweep" methods. External simulators which implement the old list-based APIs will still function properly (through the magic of `@value.alternative`), but these simulators are advised to switch to the iterator-based API for better performance.

Note that the list-based API is _not_ deprecated, and we have no plans to remove it at this time.