Simultaneous readout (#4415)
Simultaneous readout, attempt 2.  The previous attempt had an error where the trial_result (trial_idx) was not being iterated correctly with multiple batches, so it was re-reading the first batch of results over and over.  This manifested in test flakiness, since it would happen if all results in the batch were the same digit (ie. about 10*2* 1/2^10 ~= 2% of the time).

Adds an experiment for simultaneous (parallel) readout.

This experiment initializes a number of bitstrings(trials) and initializes the qubits to those bitstrings and immediately measures them It then estimates zero state and one state errors averaged over all of the trials where the given qubit was initialized to |0> or |1> respectively.