Improve speed of random circuit generation (take 2) (#4430)
Fixes #4337 (again)

Changes equality check to identity check to give the perf improvement. Since the gate instances are reused, and since equality checking if XZPowGate is very expensive, we change equality check to identity check and get a big perf boost. This issue was originally fixed in a different way in #4428, but that has been reverted in llieu of this fix which is simpler, faster, and not a breaking change.

ref https://github.com/quantumlib/Cirq/pull/4428#issuecomment-898942138

@mpharrigan