Use duet for asynchrony, in particular for Collector (#4009)
This switches async sampler methods to use the duet library (https://github.com/google/duet). Duet is "reentrant" in that it supports nested invocations of the event loop (duet.run). This differs from most other python async libraries which are not natively reentrant (without patching as in the case of nest-asyncio for asyncio). Re-entrancy makes it possible to refactor code incrementally to use duet whereas non-reentrant libraries typically require more global changes.

The asyncio-based work pool and testing utilities are deprecated and will be removed in a future release.

Review: @mpharrigan, @dstrain115