[cirqflow] Quantum Executable (#4527)
An executable quantum program.

This serves a similar purpose to `cirq.Circuit` with some key differences. First, a quantum
executable contains all the relevant context for execution including parameters as well as
the desired number of repetitions. Second, this object is immutable. Finally, there are
optional fields enabling a higher level of abstraction for certain aspects of the executable.

This is part 1 of a series of data containers for structuring experiment workflows. Diagrammatically, the whole plan looks something like this:

![RequestResponse2 (1)](https://user-images.githubusercontent.com/4967059/134717905-f03b3884-e0da-4221-841d-cccec0137193.png)

@MichaelBroughton to review
@mrwojtek cc for visibility