Add metadata property to device. (#4868)
Adds metadata property default to None to all devices. I'm going to leave this open to vendors to add to this as they see fit. Part of #4743 . Note that if in the process of deprecating other things for #4744 , I will pull vendor devices onto using metadata to keep behavior working.