Fix default QuantumStateRepresentation sampler for qudits (#5234)
Now that there's a default sample method for QuantumStateRepr, I updated a test function to use that instead (by removing the custom implementation that test used previously). The test below started to fail (it returned `[[1], [1]]` even though all breakpoints showed that the measurement was still `[[2], [2]]`, which was *super* confusing). I finally tracked it down to the dtype being `bool` in that default `sample` method. Changed that to uint8, and now all tests pass using the default sampler. I didn't add any new tests because the changes to the existing test validates it.

https://github.com/quantumlib/Cirq/blob/1522b024d2613b9acd0a57ca9eafa58371063484/cirq-core/cirq/sim/simulator_base_test.py#L237-L240

Also piggybacked a small todo from one of my other tests into this PR.