Speed up circuit building by not forgetting about cached objects (#5280)
This speeds up a 10^2 qubit by 10^3 deep creation of a circuit made up entirely of `cirq.X` gates, created by just appending these gates onto the circuit, from 25s to 5s.

The issue is that Moment's are immutable, so they need to be copied when adding in new operations.  Before this PR we don't copy two cached objects, the measurement key objects, and control keys during this.  This copies these caches over and update them.

Because using `insert` on circuit for earliest insertion strategy has to look up measurement keys or control keys (in order to not move an object with such a key before a moment that has such a key), moments during creating are always being asked what their measurement and control keys are.