Fix tests that break python-repeat (#5431)
I use python-repeat to check for flaky tests.  It allows you to pass in `--count XXX` to repeat a test multiple times.  Interestingly, for parameterized tests the decorator that creates the parameterized tests is run once globally, and then the tests are repeated.  This means that these tests can catch places where the parameterized values are mutated in some way during a test.  

This fixes the cases found from `check/pytest --count 100`.  In general mutating parameters called to functions without being very explicit about it is something we should try to avoid.